requirements.txt
confluent-kafka
fastapi
uvicorn
requests
polyline
